+++ src/mixins/hasMultioptions.js
@@ -1,5 +1,19 @@
+<<<<<<< HEAD
+=======
+import flatten from 'lodash/flatten'
+import get from 'lodash/get'
+
+>>>>>>> cea04db... The multioption components now allows you to select which attribute should be used as as value and text in the options
 const hasMultioptions = {
   props: {
+    valueAttribute: {
+      type: String,
+      default: null
+    },
+    textAttribute: {
+      type: String,
+      default: null
+    },
     options: {
       type: [Array, Object],
       default: () => []
@@ -18,6 +32,18 @@ const hasMultioptions = {
     },
   },
   methods: {
+    guessOptionValue(option) {
+      if (this.valueAttribute) {
+        return get(option, this.valueAttribute) 
+      }
+      return get(option, 'value', get(option, 'id', get(option, 'text')))
+    },
+    guessOptionText(option) {
+      if (this.textAttribute) {
+        return get(option, this.textAttribute) 
+      }
+      return get(option, 'text', get(option, 'label'))
+    },
     normalizeOption (option) {
       if (['string', 'number', 'boolean'].includes(typeof option)) {
         return {
@@ -26,29 +52,9 @@ const hasMultioptions = {
         }
       }
 
-      const value = option.hasOwnProperty('value')
-        ? option.value
-        : (
-          option.hasOwnProperty('id')
-            ? option.id
-            : (
-              option.hasOwnProperty('text')
-                ? option.text
-                : null
-              )
-        )
-
-      const text = option.hasOwnProperty('text')
-        ? option.text
-        : (
-          option.hasOwnProperty('label')
-            ? option.label
-            : null
-        )
-
       return {
-        value,
-        text
+        value: this.guessOptionValue(option),
+        text: this.guessOptionText(option),
       }
     },
   }