+++ tests/unit/TRadioGroup.spec.js
@@ -124,6 +124,26 @@ describe('TRadioGroup.vue', () => {
     expect(wrapper.vm.normalizedOptions).toEqual(expectedOptions);
   })
 
+  it('accept dynamic value/text attributes', () => {
+    const objectsWithCustomAttribs = [
+      { key: 'A', description: 'A' },
+      { key: 'B', description: 'B' },
+      { key: 'C', description: 'C' },
+    ]
+    
+    const expectedOptions = objectsWithCustomAttribs.map(option => ({ value: option.key, text: option.description }))
+    
+    const wrapper = shallowMount(TRadioGroup, {
+      propsData: {
+        options: objectsWithCustomAttribs,
+        valueAttribute: 'key',
+        textAttribute: 'description',
+      }
+    })
+
+    expect(wrapper.vm.normalizedOptions).toEqual(expectedOptions);
+  })
+
   it('disables the radio inputs', () => {
     const options = ['Option A', 'Option B', 'Option C']
     const wrapper = mount(TRadioGroup, {